+++ Demo/Scripts/DebugLogger.cs
@@ -1,0 +1,9 @@
+using UnityEngine;
+
+namespace UStrap.Demo
+{
+    public class DebugLogger : MonoBehaviour
+    {
+        public void Log(string message) => Debug.Log(message);
+    }
+}