+++ Scripts/Editor/MenuItems.cs
@@ -1,0 +1,7 @@
+namespace UStrap.Editor
+{
+    public static class MenuItems
+    {
+        //TODO
+    }
+}