+++ CHANGELOG.md
@@ -1,3 +1,11 @@
-[Version 0.2 | November 27, 2022]
+[Version 0.2 | November 29, 2022]
 • Created Unity package for UStrap
+• Updating USS style to use Bootstrap 5
 • Created new ExtendedVisualElement in Extensions
+• Created new ExtendedStrings in Extensions
+• Created new UIElementClickEvent abstract class in Data
+• Created new SliderClickEvent wrapper class in Data
+• Created new MenuItems static helper in Editor
+• Refactored DOMManager and added new functions for Showing/Hiding elements
+• Created new Demo scenes for Buttons and Height & Width
+• Created new DebugLogger script in Demo